Source as of beta 21 and some fixes
git-svn-id: https://ja2svn.mooo.com/source/ja2@1502 3b4a5df2-a311-0410-b5c6-a8a6f20db521

+++ branches/Overhaul/SCREENS.cpp
@@ -1,4 +1,3 @@
-// WANNE 2 <changed some lines>
 #ifdef PRECOMPILEDHEADERS
 	#include "JA2 All.h"
 #else
@@ -20,13 +19,12 @@ int iScreenMode = 0;	/* 0 for Fullscreen */
 #endif
 BOOL bScreenModeCmdLine = FALSE;
 
-int SCREEN_WIDTH;
-int SCREEN_HEIGHT;
+UINT16 SCREEN_WIDTH;
+UINT16 SCREEN_HEIGHT;
 
-int iScreenWidthOffset;
-int iScreenHeightOffset;
+INT16 iScreenWidthOffset;
+INT16 iScreenHeightOffset;
 
-// WANNE 2
 BOOLEAN fDisplayOverheadMap;
 
 Screens GameScreens[MAX_SCREENS] =
